using datadef for LFV only
