+++ MetaData/scripts/get_mc_lumi.py
@@ -38,7 +38,7 @@ if __name__ == "__main__":
         import FinalStateAnalysis.MetaData.data8TeVNew as datadefs
     elif args.sqrts == 13:
         sys.stderr.write("Using 13 TeV data definitions\n")
-        import FinalStateAnalysis.MetaData.data13TeV as datadefs
+        import FinalStateAnalysis.MetaData.data13TeV_LFV as datadefs
 
     sample_xsec = datadefs.datadefs[args.sample]['x_sec']/picobarns
